Redirect root / to /select_account and centralise redirect logic there
